backend/requirements.txt
Django
djangorestframework
django-cors-headers
django-environ
psycopg2-binary
gunicorn
pytest

# to profile api calls, there is probably a better solution to this using prometheus and grafana?
drf-api-logger 

# Caching server
django-redis
# DB Caching
django-cachalot

# A nice helper to get nested routers 
drf-nested-routers

# Django helpers
django-extensions

# WebSockets
channels
channels_redis
daphne